various small fixes

+++ chaossource/g_local.h
@@ -1,6 +1,7 @@
 // g_local.h -- local definitions for game module
 
 #include "q_shared.h"
+#include <stdint.h>
 
 // define GAME_INCLUDE so that game.h does not define the
 // short, server-visible gclient_t and edict_t structures,
@@ -533,10 +534,10 @@ extern	int	meansOfDeath;
 
 extern	edict_t			*g_edicts;
 
-#define	FOFS(x) (long int)&(((edict_t *)0)->x)
-#define	STOFS(x) (long int)&(((spawn_temp_t *)0)->x)
-#define	LLOFS(x) (long int)&(((level_locals_t *)0)->x)
-#define	CLOFS(x) (long int)&(((gclient_t *)0)->x)
+#define FOFS(x) (intptr_t)&(((edict_t *)0)->x)
+#define	STOFS(x) (intptr_t)&(((spawn_temp_t *)0)->x)
+#define	LLOFS(x) (intptr_t)&(((level_locals_t *)0)->x)
+#define	CLOFS(x) (intptr_t)&(((gclient_t *)0)->x)
 
 #define random()	((rand () & 0x7fff) / ((float)0x7fff))
 #define crandom()	(2.0 * (random() - 0.5))
